fix(gpg-cli): tests should support windows
