+++ pkg/backend/crypto/gpg/cli/generate_windows_test.go
@@ -1,0 +1,18 @@
+package cli
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCreatePrivateKey(t *testing.T) {
+	ctx := context.Background()
+
+	g := &GPG{}
+	g.binary = "rundll32"
+
+	assert.NoError(t, g.CreatePrivateKeyBatch(ctx, "foo", "foo@bar.com", "bar"))
+	assert.NoError(t, g.CreatePrivateKey(ctx))
+}
