+++ pkg/backend/crypto/gpg/cli/gpg_test.go
@@ -34,29 +34,3 @@ func TestGPG(t *testing.T) {
 	assert.Equal(t, "gpg", g.Ext())
 	assert.Equal(t, ".gpg-id", g.IDFile())
 }
-
-func TestDetectBinaryCandidates(t *testing.T) {
-	bins, err := detectBinaryCandidates("foobar")
-	assert.NoError(t, err)
-	assert.Equal(t, []string{"gpg2", "gpg1", "gpg", "foobar"}, bins)
-}
-
-func TestEncrypt(t *testing.T) {
-	ctx := context.Background()
-
-	g := &GPG{}
-	g.binary = "true"
-
-	_, err := g.Encrypt(ctx, []byte("foo"), nil)
-	assert.NoError(t, err)
-}
-
-func TestDecrypt(t *testing.T) {
-	ctx := context.Background()
-
-	g := &GPG{}
-	g.binary = "true"
-
-	_, err := g.Decrypt(ctx, []byte("foo"))
-	assert.NoError(t, err)
-}
